refactor: 멀티 모듈 구조 적용
