+++ game-server/src/main/java/com/example/gameserver/common/GameCommandHandler.java
@@ -1,0 +1,8 @@
+
+package com.example.gameserver.common;
+
+import com.example.gameserver.handler.GameCommand;
+
+public interface GameCommandHandler {
+    GameCommand getCommand();
+}
