+++ game-server/src/main/java/com/example/gameserver/handler/GameCommand.java
@@ -1,0 +1,9 @@
+package com.example.gameserver.handler;
+
+public enum GameCommand {
+    JOIN_QUEUE,
+    LEAVE_QUEUE,
+    PLAY,
+    SURRENDER,
+    EXIT;
+}
