+++ game-server/src/main/java/com/example/gameserver/handler/ResponseType.java
@@ -1,0 +1,11 @@
+package com.example.gameserver.handler;
+
+public enum ResponseType {
+    CONNECTED,
+    WAITING,
+    GAME_START,
+    GAME_RESULT,
+    GAME_END,
+    ERROR,
+    OPPONENT_DISCONNECTED
+}
