+++ game-server/src/main/java/com/example/gameserver/handler/SimpleCommandHandler.java
@@ -1,0 +1,8 @@
+package com.example.gameserver.handler;
+
+import com.example.gameserver.common.GameCommandHandler;
+import org.springframework.web.reactive.socket.WebSocketSession;
+
+public interface SimpleCommandHandler extends GameCommandHandler {
+    void handle(WebSocketSession session);
+}
