+++ game-server/src/main/java/com/example/gameserver/model/GameChoice.java
@@ -1,0 +1,5 @@
+package com.example.gameserver.model;
+
+public enum GameChoice {
+    ROCK, PAPER, SCISSORS
+}
