+++ game-server/src/main/java/com/example/gameserver/model/GameMessage.java
@@ -1,0 +1,10 @@
+package com.example.gameserver.model;
+
+import com.example.gameserver.handler.GameCommand;
+import lombok.Data;
+
+@Data
+public class GameMessage<T> {
+    private GameCommand command;
+    private T data;
+}
