+++ game-server/src/main/java/com/example/gameserver/model/GameResponse.java
@@ -1,0 +1,13 @@
+package com.example.gameserver.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class GameResponse {
+    private String type;
+    private String message;
+}
