+++ game-server/src/main/java/com/example/gameserver/model/GameResult.java
@@ -1,0 +1,14 @@
+package com.example.gameserver.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class GameResult {
+    private String winnerId;
+    private String loserId;
+    private GameChoice winnerChoice;
+    private GameChoice loserChoice;
+    private boolean isDraw;
+}
