+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'online-game'
+
+include 'game-server'
+include 'payment-server'
