--- src/main/java/com/example/onlinegame/OnlineGameApplication.java
@@ -1,13 +1,0 @@
-package com.example.onlinegame;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnlineGameApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OnlineGameApplication.class, args);
-    }
-
-}
