--- src/main/java/com/example/onlinegame/common/GameCommandHandler.java
@@ -1,8 +1,0 @@
-
-package com.example.onlinegame.common;
-
-import com.example.onlinegame.handler.GameCommand;
-
-public interface GameCommandHandler {
-    GameCommand getCommand();
-}
