--- src/main/java/com/example/onlinegame/handler/GameCommand.java
@@ -1,9 +1,0 @@
-package com.example.onlinegame.handler;
-
-public enum GameCommand {
-    JOIN_QUEUE,
-    LEAVE_QUEUE,
-    PLAY,
-    SURRENDER,
-    EXIT;
-}
