--- src/main/java/com/example/onlinegame/handler/ResponseType.java
@@ -1,11 +1,0 @@
-package com.example.onlinegame.handler;
-
-public enum ResponseType {
-    CONNECTED,
-    WAITING,
-    GAME_START,
-    GAME_RESULT,
-    GAME_END,
-    ERROR,
-    OPPONENT_DISCONNECTED
-}
