--- src/main/java/com/example/onlinegame/handler/SimpleCommandHandler.java
@@ -1,8 +1,0 @@
-package com.example.onlinegame.handler;
-
-import com.example.onlinegame.common.GameCommandHandler;
-import org.springframework.web.reactive.socket.WebSocketSession;
-
-public interface SimpleCommandHandler extends GameCommandHandler {
-    void handle(WebSocketSession session);
-}
