--- src/main/java/com/example/onlinegame/model/GameChoice.java
@@ -1,5 +1,0 @@
-package com.example.onlinegame.model;
-
-public enum GameChoice {
-    ROCK, PAPER, SCISSORS
-}
