--- src/main/java/com/example/onlinegame/model/GameMessage.java
@@ -1,10 +1,0 @@
-package com.example.onlinegame.model;
-
-import com.example.onlinegame.handler.GameCommand;
-import lombok.Data;
-
-@Data
-public class GameMessage<T> {
-    private GameCommand command;
-    private T data;
-}
