--- src/main/java/com/example/onlinegame/model/GameResponse.java
@@ -1,13 +1,0 @@
-package com.example.onlinegame.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class GameResponse {
-    private String type;
-    private String message;
-}
