--- src/test/java/com/example/onlinegame/OnlineGameApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.onlinegame;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OnlineGameApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
